// lib/types.ts
export type PainPoint = {
  title: string;
  description: string;
};

export type Solution = {
  title: string;
  description: string;
};

export type PortfolioItem = {
  title: string;
  description: string;
  category: string;
  repo: string;
  githubUrl: string;
  demoUrl: string;
};

export type ProcessStep = {
  step: string;
  title: string;
  description: string;
};